Refactor compile.ts, move test files to their own place, upgrade packages.
compile.ts kinda works, but I don't like the way it's handling the ast. Also,
the transform api doesn't have access to calculated types. For that, I need
to use the type checker. redux-test.ts is doing things differently and I
prefer that approach: the ast handler is easy to read, and it's using the type
checker, but for some reason the type checker child visitor function is not
going through all the entities in a file, and it's not going down the whole tree,
only the first level. I have no idea why.